==> 01. sdoh repository development 20250115.Rmd <==
---
title: "SDoH Code"
author: "Carina Korcel"
date: "2024-12-18"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r load libraries}
library(dplyr)
library(censusapi)
library(readxl)
library(tidyr)
library(ggplot2)
library(httr)
library(tidyverse)
```

##########################################################################################################
########################################## Zip Code Level Table ##########################################
##########################################################################################################

# Data Sources

  1. American Community Survey (ACS, 2021) - Survey data from 3.5 million addresses on 40+ topics conducted by the Census Bureau. Topics include age, children, veterans, commuting, education, income, and employment. The data used here are 5-year estimates. New data is available every year. Data reflects 5-year period not a point in time. Data is collected through internet, mail, and in-person visits.
  
  2. CDC Local Data for Better Health (CDC, 2021) - Estimates for 36 health measures designed to be used to identify emerging health problems and to help develop and carry out effective, targeted public health prevention activities.
  
  3. CDC Social Vulnerability Index (SVI, 2018) - Social vulnerability is the degree to which a community exhibits certain social conditions that may affect that community's ability to prevent human suffering and financial loss in the event of a disaster. Index indicating the relative vulnerability of every U.S. ZCTA. Ranks on 5 social factors and groups them into 4 related themes.

# Data Pulling
  1. ACS - via api call to Census Bureau
  2. CDC - via excel file downloaded from CDC website
  3. SVI - via excel file downloaded from CDC website

###############################################################
############### American Community Survey (ACS) ###############
###############################################################

# Import ACS data
```{r pull ACS data from Census Bureau website}
# establish Census API connection
Sys.setenv(CENSUS_KEY= "key") # DO NOT SHARE THIS
readRenviron("~/.Renviron")
Sys.getenv("CENSUS_KEY")

# build function to execute census api call
censusapicall <- function(x){
  
  # get field names
  variables <- listCensusMetadata(
    name="acs/acs5",
    vintage="2021",
    type="variables",
    group=x) %>%
    filter(grepl("EA",name)==FALSE,
           grepl("MA",name)==FALSE,
           grepl("M",name)==FALSE)
    
  # select field names and descriptions
  vars <- variables$name
  cols <- variables$label
    
  # pull selected fields
  data <- censusapi::getCensus(
    name="acs/acs5",
    vintage="2021",
    vars=c("NAME", vars),
    region="zip code tabulation area:*") 
    
  # adding column names to data
  colnames(data) <- c("ZCTA", "Name", cols)
  
  # add table to environment
  assign(x, data, envir = .GlobalEnv)
  
}

# create a data dictionary of select tables
dictionary <- as.data.frame(rbind(
  c("B01001", "population by sex and age"),
  c("B02001", "race"),
  c("B03001", "hispanic/latino"),
  c("B05001", "citizenship"),
  c("B21001", "veteran status"),
  c("B11001", "household type"),
  c("B25104", "monthly housing costs"),
  c("B23001", "employment status"),
  c("B17001", "poverty status"),
  c("B17024", "ratio of income to poverty level"),
  c("B19013", "median household income"),
  c("B19083", "Gini index of income inequality"),
  c("B22003", "food stamps/SNAP"),
  c("B15001", "educational attainment for 18+"),
  c("B08006", "means of transportation to work"),
  c("B27001", "health insurance coverage status"),
  c("B18101", "disability status"),
  c("B16005", "limited English"),
  c("B28011", "no internet")
  ) )
colnames(dictionary) <- c("table", "topic")

# use censusapicall function to pull selected tables
api_calls <- sapply(dictionary$table, censusapicall)
```

# Data Wrangling/Cleaning ACS Data
```{r clean ACS data}
#  1. aggregate the detailed cross-section data
#  2. convert totals into proportions
#  3. combine cleaned tables into 1 ACS table

# population by sex and age
B01001_clean <- B01001 %>%
  mutate(total_pop = `Estimate!!Total:`,
         prop_male = `Estimate!!Total:!!Male:`/`Estimate!!Total:`,
         prop_female = `Estimate!!Total:!!Female:`/`Estimate!!Total:`,
         total_age_under_18 = (  rowSums(B01001[,colnames(B01001)[grepl("under 5 years", colnames(B01001), ignore.case = TRUE)]]) +
                                   rowSums(B01001[,colnames(B01001)[grepl("5 to 9 years", colnames(B01001), ignore.case = TRUE)]]) +
                                   rowSums(B01001[,colnames(B01001)[grepl("10 to 14 years", colnames(B01001), ignore.case = TRUE)]]) +
                                   rowSums(B01001[,colnames(B01001)[grepl("15 to 17 years", colnames(B01001), ignore.case = TRUE)]]) ) ,
         total_age_18_to_64 = (  rowSums(B01001[,colnames(B01001)[grepl("18 and 19 years", colnames(B01001), ignore.case = TRUE)]]) +
                                   rowSums(B01001[,colnames(B01001)[grepl("20 years", colnames(B01001), ignore.case = TRUE)]]) +
                                   rowSums(B01001[,colnames(B01001)[grepl("21 years", colnames(B01001), ignore.case = TRUE)]]) +
                                   rowSums(B01001[,colnames(B01001)[grepl("22 to 24 years", colnames(B01001), ignore.case = TRUE)]]) +
                                   rowSums(B01001[,colnames(B01001)[grepl("25 to 29 years", colnames(B01001), ignore.case = TRUE)]]) +
                                   rowSums(B01001[,colnames(B01001)[grepl("30 to 34 years", colnames(B01001), ignore.case = TRUE)]]) +
                                   rowSums(B01001[,colnames(B01001)[grepl("35 to 39 years", colnames(B01001), ignore.case = TRUE)]]) +
                                   rowSums(B01001[,colnames(B01001)[grepl("40 to 44 years", colnames(B01001), ignore.case = TRUE)]]) +
                                   rowSums(B01001[,colnames(B01001)[grepl("45 to 49 years", colnames(B01001), ignore.case = TRUE)]]) +
                                   rowSums(B01001[,colnames(B01001)[grepl("50 to 54 years", colnames(B01001), ignore.case = TRUE)]]) +
                                   rowSums(B01001[,colnames(B01001)[grepl("55 to 59 years", colnames(B01001), ignore.case = TRUE)]]) +
                                   rowSums(B01001[,colnames(B01001)[grepl("60 and 61 years", colnames(B01001), ignore.case = TRUE)]]) +
                                   rowSums(B01001[,colnames(B01001)[grepl("62 to 64 years", colnames(B01001), ignore.case = TRUE)]]) ) ,
         total_age_over_64 = (  rowSums(B01001[,colnames(B01001)[grepl("65 and 66 years", colnames(B01001), ignore.case = TRUE)]]) +
                                   rowSums(B01001[,colnames(B01001)[grepl("67 to 69 years", colnames(B01001), ignore.case = TRUE)]]) +
                                   rowSums(B01001[,colnames(B01001)[grepl("70 to 74 years", colnames(B01001), ignore.case = TRUE)]]) +
                                   rowSums(B01001[,colnames(B01001)[grepl("75 to 79 years", colnames(B01001), ignore.case = TRUE)]]) +
                                   rowSums(B01001[,colnames(B01001)[grepl("80 to 84 years", colnames(B01001), ignore.case = TRUE)]]) +
                                   rowSums(B01001[,colnames(B01001)[grepl("85 years and over", colnames(B01001), ignore.case = TRUE)]]) ),
         prop_age_under_18 = total_age_under_18/`Estimate!!Total:`,
         prop_age_18_to_64 = total_age_18_to_64/`Estimate!!Total:`,
         prop_age_over_64 = total_age_over_64/`Estimate!!Total:` ) %>%
  select(ZCTA,
         prop_female,
         prop_age_under_18,
         prop_age_18_to_64,
         prop_age_over_64)

# race
B02001_clean <- B02001 %>%
  mutate(total_race = `Estimate!!Total:`,
         prop_race_two_or_more = `Estimate!!Total:!!Two or more races:`/total_race,
         prop_race_alone_other = `Estimate!!Total:!!Some other race alone`/total_race,
         prop_race_alone_white = `Estimate!!Total:!!White alone`/total_race,
         prop_race_alone_HI_PI = `Estimate!!Total:!!Native Hawaiian and Other Pacific Islander alone`/total_race,
         prop_race_alone_asian = `Estimate!!Total:!!Asian alone`/total_race,
         prop_race_alone_AI_AK = `Estimate!!Total:!!American Indian and Alaska Native alone`/total_race,
         prop_race_alone_black = `Estimate!!Total:!!Black or African American alone`/total_race,
         prop_minority = (prop_race_two_or_more +
                          prop_race_alone_other +
                          prop_race_alone_HI_PI +
                          prop_race_alone_asian +
                          prop_race_alone_AI_AK +
                          prop_race_alone_black) ) %>%
  select(ZCTA,
         prop_race_two_or_more,
         prop_race_alone_other,
         prop_race_alone_white,
         prop_race_alone_HI_PI,
         prop_race_alone_asian,
         prop_race_alone_AI_AK,
         prop_race_alone_black,
         prop_minority)

# hispanic/latino
B03001_clean <-  B03001 %>%
  mutate(prop_hispanic_latino = `Estimate!!Total:!!Hispanic or Latino:`/`Estimate!!Total:`) %>%
  select(ZCTA,
         prop_hispanic_latino )

# nativity and citizenship status
B05001_clean <- B05001 %>%
  mutate(total=`Estimate!!Total:`,
         prop_not_citizen = `Estimate!!Total:!!Not a U.S. citizen`/total,
         prop_citizen = 1-prop_not_citizen) %>%
  select(ZCTA,
         prop_citizen)

# veteran status
B21001_clean <- B21001 %>%
  mutate(prop_veteran = `Estimate!!Total:!!Veteran`/`Estimate!!Total:`) %>%
  select(ZCTA,
         prop_veteran)

# household type
B11001_clean <- B11001 %>%
  mutate(prop_family_HH = `Estimate!!Total:!!Family households:`/`Estimate!!Total:`,
         prop_nonfamily_HH = `Estimate!!Total:!!Nonfamily households:`/`Estimate!!Total:`) %>%
  select(ZCTA,
         prop_family_HH)

# monthly housing costs
B25104_clean <- B25104 %>%
  mutate(prop_housing_no_cash_rent = `Estimate!!Total:!!No cash rent`/`Estimate!!Total:`,
         prop_housing_less_than_500 = (`Estimate!!Total:!!Less than $100`+
                                         `Estimate!!Total:!!$100 to $199`+
                                         `Estimate!!Total:!!$200 to $299`+
                                         `Estimate!!Total:!!$300 to $399`+
                                         `Estimate!!Total:!!$400 to $499`)/`Estimate!!Total:`,
         prop_housing_between_500_and_999 = (`Estimate!!Total:!!$500 to $599`+
                                               `Estimate!!Total:!!$600 to $699`+
                                               `Estimate!!Total:!!$700 to $799`+
                                               `Estimate!!Total:!!$800 to $899`+
                                               `Estimate!!Total:!!$900 to $999`)/`Estimate!!Total:`,
         prop_housing_between_1000_and_1500 = `Estimate!!Total:!!$1,000 to $1,499`/`Estimate!!Total:`,
         prop_housing_between_1500_and_2000 = `Estimate!!Total:!!$1,500 to $1,999`/`Estimate!!Total:`,
         prop_housing_between_2000_and_2500 = `Estimate!!Total:!!$2,000 to $2,499`/`Estimate!!Total:`,
         prop_housing_between_2500_and_3000 = `Estimate!!Total:!!$2,500 to $2,999`/`Estimate!!Total:`,
         prop_housing_greater_than_3000 = `Estimate!!Total:!!$3,000 or more`/`Estimate!!Total:`) %>%
  select(ZCTA,
         prop_housing_no_cash_rent,
         prop_housing_less_than_500,
         prop_housing_between_500_and_999,
         prop_housing_between_1000_and_1500,
         prop_housing_between_1500_and_2000,
         prop_housing_between_2000_and_2500,
         prop_housing_between_2500_and_3000,
         prop_housing_greater_than_3000)

# employment status 
B23001_clean <- B23001 %>%
  mutate(total_armed_forces = rowSums(B23001[,colnames(B23001)[grepl("in armed forces", colnames(B23001), ignore.case = TRUE)]]),
         total_employed = rowSums(B23001[,colnames(B23001)[grepl("Employed", colnames(B23001), ignore.case = FALSE)]]),
         total_unemployed = rowSums(B23001[,colnames(B23001)[grepl("unemployed", colnames(B23001), ignore.case = TRUE)]]),
         total_in_labor_force = total_employed + total_unemployed,
         total_not_in_labor_force = rowSums(B23001[,colnames(B23001)[grepl("not in labor force", colnames(B23001), ignore.case = TRUE)]])) %>%
  mutate(prop_armed_forces = total_armed_forces/`Estimate!!Total:`,
         prop_employed = total_employed/`Estimate!!Total:`,
         prop_unemployed = total_unemployed/`Estimate!!Total:`,
         prop_not_in_labor_force = total_not_in_labor_force/`Estimate!!Total:`) %>%
  select(ZCTA,
         prop_armed_forces,
         prop_employed,
         prop_unemployed,
         prop_not_in_labor_force)

# poverty status
B17001_clean <- B17001 %>%
  mutate(prop_at_or_above_poverty_level = `Estimate!!Total:!!Income in the past 12 months at or above poverty level:`/`Estimate!!Total:`,
         prop_below_poverty_level = `Estimate!!Total:!!Income in the past 12 months below poverty level:`/`Estimate!!Total:`) %>%
  select(ZCTA,
         prop_below_poverty_level)

# ratio of income to poverty level in past 12 months
B17024_clean <- B17024 %>%
  mutate(total_poverty_ratio_5.00_and_over = rowSums(B17024[,colnames(B17024)[grepl("5.00 and over", colnames(B17024), ignore.case = TRUE)]]),
         total_poverty_ratio_4.00_to_4.99 =  rowSums(B17024[,colnames(B17024)[grepl("4.00 to 4.99", colnames(B17024), ignore.case = TRUE)]]),
         total_poverty_ratio_3.00_to_3.99 =  rowSums(B17024[,colnames(B17024)[grepl("3.00 to 3.99", colnames(B17024), ignore.case = TRUE)]]),
         total_poverty_ratio_2.00_to_2.99 =  rowSums(B17024[,colnames(B17024)[grepl("2.00 to 2.99", colnames(B17024), ignore.case = TRUE)]]),
         total_poverty_ratio_1.85_to_1.99 =  rowSums(B17024[,colnames(B17024)[grepl("1.85 to 1.99", colnames(B17024), ignore.case = TRUE)]]),
         total_poverty_ratio_1.75_to_1.84 =  rowSums(B17024[,colnames(B17024)[grepl("1.75 to 1.84", colnames(B17024), ignore.case = TRUE)]]),
         total_poverty_ratio_1.50_to_1.74 =  rowSums(B17024[,colnames(B17024)[grepl("1.50 to 1.74", colnames(B17024), ignore.case = TRUE)]]),
         total_poverty_ratio_1.25_to_1.49 =  rowSums(B17024[,colnames(B17024)[grepl("1.25 to 1.49", colnames(B17024), ignore.case = TRUE)]]),
         total_poverty_ratio_1.00_to_1.24 =  rowSums(B17024[,colnames(B17024)[grepl("1.00 to 1.24", colnames(B17024), ignore.case = TRUE)]]),
         total_poverty_ratio_0.75_to_0.99 =  rowSums(B17024[,colnames(B17024)[grepl(".75 to .99", colnames(B17024), ignore.case = TRUE)]]),
         total_poverty_ratio_0.50_to_0.74 =  rowSums(B17024[,colnames(B17024)[grepl(".50 to .74", colnames(B17024), ignore.case = TRUE)]]),
         total_poverty_ratio_under_0.50 =  rowSums(B17024[,colnames(B17024)[grepl("under .50", colnames(B17024), ignore.case = TRUE)]])
  ) %>%
  mutate(prop_poverty_ratio_5.00_and_over = total_poverty_ratio_5.00_and_over / `Estimate!!Total:`,
         prop_poverty_ratio_4.00_to_4.99 = total_poverty_ratio_4.00_to_4.99 / `Estimate!!Total:`,
         prop_poverty_ratio_3.00_to_3.99 = total_poverty_ratio_3.00_to_3.99 / `Estimate!!Total:`,
         prop_poverty_ratio_2.00_to_2.99 = total_poverty_ratio_2.00_to_2.99 / `Estimate!!Total:`,
         prop_poverty_ratio_1.85_to_1.99 = total_poverty_ratio_1.85_to_1.99 / `Estimate!!Total:`,
         prop_poverty_ratio_1.75_to_1.84 = total_poverty_ratio_1.75_to_1.84 / `Estimate!!Total:`,
         prop_poverty_ratio_1.50_to_1.74 = total_poverty_ratio_1.50_to_1.74 / `Estimate!!Total:`,
         prop_poverty_ratio_1.25_to_1.49 = total_poverty_ratio_1.25_to_1.49 / `Estimate!!Total:`,
         prop_poverty_ratio_1.00_to_1.24 = total_poverty_ratio_1.00_to_1.24 / `Estimate!!Total:`,
         prop_poverty_ratio_0.75_to_0.99 = total_poverty_ratio_0.75_to_0.99 / `Estimate!!Total:`,
         prop_poverty_ratio_0.50_to_0.74 = total_poverty_ratio_0.50_to_0.74 / `Estimate!!Total:`,
         prop_poverty_ratio_under_0.50 = total_poverty_ratio_under_0.50 / `Estimate!!Total:`) %>%
  mutate(prop_poverty_ratio_under_1.50 = prop_poverty_ratio_1.25_to_1.49 +
                                         prop_poverty_ratio_1.00_to_1.24 +
                                         prop_poverty_ratio_0.75_to_0.99 +
                                         prop_poverty_ratio_0.50_to_0.74 +
                                         prop_poverty_ratio_under_0.50) %>%
  select(ZCTA,
         prop_poverty_ratio_under_1.50)

# median household income in the past 12 months (2021 inflation adjusted dollars)
B19013_clean <- B19013 %>%
  mutate(median_HH_income = `Estimate!!Median household income in the past 12 months (in 2021 inflation-adjusted dollars)`,
         median_HH_income = ifelse(median_HH_income < 0, NA, median_HH_income))%>%
  select(ZCTA,
         median_HH_income)

# Gini index of income inequality
B19083_clean <- B19083 %>%
  mutate(gini_index = `Estimate!!Gini Index`,
         gini_index = ifelse(gini_index < 0, NA, gini_index)) %>%
  select(ZCTA,
         gini_index)

# received food stamps/SNAP in the past 12 months for households
B22003_clean <- B22003 %>%
  mutate(prop_food_stamps = `Estimate!!Total:!!Household received Food Stamps/SNAP in the past 12 months:`/`Estimate!!Total:`,
         prop_no_food_stamps= `Estimate!!Total:!!Household did not receive Food Stamps/SNAP in the past 12 months:`/`Estimate!!Total:`) %>%
  select(ZCTA,
         prop_food_stamps)

# educational attainment for 18+
B15001_clean <- B15001 %>%
  mutate(total_no_high_school = rowSums(B15001[,colnames(B15001)[grepl("less than 9th grade", colnames(B15001), ignore.case = TRUE)]]),
         total_some_high_school = rowSums(B15001[,colnames(B15001)[grepl("9th to 12th grade", colnames(B15001), ignore.case = TRUE)]]),
         total_high_school_grad = rowSums(B15001[,colnames(B15001)[grepl("high school", colnames(B15001), ignore.case = TRUE)]]),
         total_some_college = rowSums(B15001[,colnames(B15001)[grepl("some college", colnames(B15001), ignore.case = TRUE)]]),
         total_associates_degree = rowSums(B15001[,colnames(B15001)[grepl("associate's degree", colnames(B15001), ignore.case = TRUE)]]),
         total_bachelors_degree = rowSums(B15001[,colnames(B15001)[grepl("bachelor's degree", colnames(B15001), ignore.case = TRUE)]]),
         total_graduate_degree_or_above = rowSums(B15001[,colnames(B15001)[grepl("graduate or professional", colnames(B15001), ignore.case = TRUE)]])) %>%
  mutate(prop_no_high_school = total_no_high_school/`Estimate!!Total:`,
         prop_some_high_school = total_some_high_school/`Estimate!!Total:`,
         prop_high_school_grad = total_high_school_grad/`Estimate!!Total:`,
         prop_some_college = total_some_college/`Estimate!!Total:`,
         prop_associates_degree = total_associates_degree/`Estimate!!Total:`,
         prop_bachelors_degree = total_bachelors_degree/`Estimate!!Total:`,
         prop_graduate_degree_or_above = total_graduate_degree_or_above/`Estimate!!Total:`) %>%
  mutate(prop_no_HS_diploma = prop_no_high_school + prop_some_high_school) %>%
  select(ZCTA,
         prop_no_HS_diploma,
         prop_high_school_grad,
         prop_some_college,
         prop_associates_degree,
         prop_bachelors_degree,
         prop_graduate_degree_or_above)
  
# means of transportation to work
B08006_clean <- B08006 %>%
  mutate(prop_car_truck_van = `Estimate!!Total:!!Car, truck, or van:`/`Estimate!!Total:`,
         prop_public_trans = `Estimate!!Total:!!Public transportation (excluding taxicab):`/`Estimate!!Total:`,
         prop_taxi_motorcycle_other = `Estimate!!Total:!!Taxicab, motorcycle, or other means`/`Estimate!!Total:`,
         prop_bicycle = `Estimate!!Total:!!Bicycle`/`Estimate!!Total:`,
         prop_walk = `Estimate!!Total:!!Walked`/`Estimate!!Total:`,
         prop_WFH = `Estimate!!Total:!!Worked from home`/`Estimate!!Total:`) %>%
  select(ZCTA,
         prop_car_truck_van,
         prop_public_trans,
         prop_taxi_motorcycle_other,
         prop_bicycle,
         prop_walk,
         prop_WFH)

# health insurance coverage status
B27001_clean <- B27001 %>%
  mutate(total_health_insurance = rowSums(B27001[,colnames(B27001)[grepl("with health insurance", colnames(B27001), ignore.case = TRUE)]]),
         total_no_health_insurance = rowSums(B27001[,colnames(B27001)[grepl("no health insurance", colnames(B27001), ignore.case = TRUE)]])) %>%
  mutate(prop_health_insurance = total_health_insurance/`Estimate!!Total:`,
         prop_no_health_insurance = total_no_health_insurance/`Estimate!!Total:`) %>%
  select(ZCTA,
         prop_health_insurance)

# disability status
B18101_clean <- B18101 %>%
  mutate(total_disability = rowSums(B18101[,colnames(B18101)[grepl("no disability", colnames(B18101), ignore.case = TRUE)]]),
         total_no_disability = rowSums(B18101[,colnames(B18101)[grepl("with a disability", colnames(B18101), ignore.case = TRUE)]])) %>%
  mutate(prop_disability = total_disability / `Estimate!!Total:`,
         prop_no_disability = total_no_disability / `Estimate!!Total:`) %>%
  select(ZCTA,
         prop_disability)

# limited English
B16005_clean <- B16005 %>%
  mutate(prop_limited_eng = (`Estimate!!Total:!!Foreign born:!!Speak other Indo-European languages:!!Speak English "not at all"`+
                               `Estimate!!Total:!!Foreign born:!!Speak Asian and Pacific Island languages:!!Speak English "not at all"` +
                               `Estimate!!Total:!!Native:!!Speak other Indo-European languages:!!Speak English "not at all"` +
                               `Estimate!!Total:!!Native:!!Speak Asian and Pacific Island languages:!!Speak English "not at all"` +
                               `Estimate!!Total:!!Foreign born:!!Speak other languages:!!Speak English "not at all"` +
                               `Estimate!!Total:!!Native:!!Speak other languages:!!Speak English "not at all"` +
                               `Estimate!!Total:!!Foreign born:!!Speak other Indo-European languages:!!Speak English "not well"` +
                               `Estimate!!Total:!!Foreign born:!!Speak Spanish:!!Speak English "not at all"` +
                               `Estimate!!Total:!!Foreign born:!!Speak Asian and Pacific Island languages:!!Speak English "not well"` +
                               `Estimate!!Total:!!Native:!!Speak other Indo-European languages:!!Speak English "not well"` +
                               `Estimate!!Total:!!Native:!!Speak Spanish:!!Speak English "not at all"` +
                               `Estimate!!Total:!!Native:!!Speak Asian and Pacific Island languages:!!Speak English "not well"` +
                               `Estimate!!Total:!!Foreign born:!!Speak Spanish:!!Speak English "not well"`+
                               `Estimate!!Total:!!Foreign born:!!Speak other languages:!!Speak English "not at all"` +
                               `Estimate!!Total:!!Native:!!Speak Spanish:!!Speak English "not well"` +
                               `Estimate!!Total:!!Native:!!Speak other languages:!!Speak English "not well"`) / `Estimate!!Total:`) %>%
  select(ZCTA,
         prop_limited_eng)

# no internet
B28011_clean <- B28011 %>%
  mutate(prop_no_internet = `Estimate!!Total:!!No Internet access`/`Estimate!!Total:`) %>%
  select(ZCTA,
         prop_no_internet)

# combine cleaned data
acs <- B01001_clean %>%
  left_join(B05001_clean, by="ZCTA") %>%
  left_join(B02001_clean, by="ZCTA") %>%
  left_join(B03001_clean, by="ZCTA") %>%
  left_join(B21001_clean, by="ZCTA") %>%
  left_join(B11001_clean, by="ZCTA") %>%
  left_join(B25104_clean, by="ZCTA") %>%
  left_join(B23001_clean, by="ZCTA") %>%
  left_join(B17001_clean, by="ZCTA") %>%
  left_join(B17024_clean, by="ZCTA") %>%
  left_join(B22003_clean, by="ZCTA") %>%
  left_join(B19013_clean, by="ZCTA") %>%
  left_join(B19083_clean, by="ZCTA") %>%
  left_join(B15001_clean, by="ZCTA") %>%
  left_join(B08006_clean, by="ZCTA") %>%
  left_join(B27001_clean, by="ZCTA") %>%
  left_join(B18101_clean, by="ZCTA") %>%
  left_join(B16005_clean, by="ZCTA") %>%
  left_join(B28011_clean, by="ZCTA") %>%
  mutate(ZCTA = as.numeric(ZCTA))

# add ACS_ to column names
colnames(acs)[-1] <- paste0("ACS_", colnames(acs)[-1])
```

##################################################################
############### Local Data for Better Health (CDC) ###############
##################################################################

# Import and clean CDC data
```{r import and clean CDC data}
# read in CDC disease burden file
cdc_file <- read.csv("F:\\Data_Science\\SDoH\\Data\\CDC_PLACES__Local_Data_for_Better_Health__ZCTA_Data_2023_release_20240103.csv")

# variable descriptions
cdc_description <- cdc_file %>%
  select(Measure,
         MeasureId) %>%
  distinct()
  
# only select crude prevalence for each disease category at ZCTA level
cdc <- cdc_file %>%
  filter(Data_Value_Type=="Crude prevalence") %>%
  mutate(ZCTA=LocationName,
         prevalence=Data_Value,
         disease_category=MeasureId) %>%
  select(ZCTA,
         prevalence,
         disease_category) %>%

  # format as a "wide" table
  tidyr::pivot_wider(names_from = disease_category,
              values_from = prevalence) %>%
  mutate(ZCTA=as.numeric(ZCTA))

# format as percentages
cdc[colnames(cdc)[-1]] <- cdc[colnames(cdc)[-1]] / 100

# add CDC_ to column names
colnames(cdc)[-1] <- paste0("CDC_", colnames(cdc)[-1])
```

####################################################################
############### CDC Social Vulnerability Index (SVI) ###############
####################################################################

# Import and clean SVI data
```{r import and clean SVI data}
svi_file <- read.csv("F:\\Data_Science\\SDoH\\Data\\SVI_2020_US_Census_Tracts.csv")
colnames(svi_file)[1] <- "ST"

svi <- svi_file %>%
  mutate(TRACT = stringr::str_pad(FIPS, 11, pad = "0")) %>%
  select(STATE,
         COUNTY,
         TRACT,
         colnames(svi_file)[grepl("EP_", colnames(svi_file), ignore.case = TRUE)]) %>%
  
  # change -999's to NA
  mutate(EP_UNEMP = ifelse(EP_UNEMP < 0, NA, EP_UNEMP),
         EP_NOHSDP = ifelse(EP_NOHSDP < 0, NA, EP_NOHSDP),
         EP_UNINSUR = ifelse(EP_UNINSUR < 0, NA, EP_UNINSUR),
         EP_AGE65 = ifelse(EP_AGE65 < 0, NA, EP_AGE65),
         EP_DISABL = ifelse(EP_DISABL < 0, NA, EP_DISABL),
         EP_MOBILE = ifelse(EP_MOBILE < 0, NA, EP_MOBILE),
         EP_NOVEH = ifelse(EP_NOVEH < 0, NA, EP_NOVEH),
         EP_AFAM = ifelse(EP_AFAM < 0, NA, EP_AFAM),
         EP_HISP = ifelse(EP_HISP < 0, NA, EP_HISP),
         EP_ASIAN = ifelse(EP_ASIAN < 0, NA, EP_ASIAN),
         EP_AIAN = ifelse(EP_AIAN < 0, NA, EP_AIAN),
         EP_NHPI = ifelse(EP_NHPI < 0, NA, EP_NHPI),
         EP_TWOMORE = ifelse(EP_TWOMORE < 0, NA, EP_TWOMORE),
         EP_OTHERRACE = ifelse(EP_OTHERRACE < 0, NA, EP_OTHERRACE))

### get zip to tract mappings (from HUDUSER)
key <- "key" # DO NOT SHARE THIS
url <- "https://www.huduser.gov/hudapi/public/usps"
response <- httr::GET(url, query = list(type = 1, query="All"), httr::add_headers(Authorization = paste("Bearer", key)))

cont <- response$content
char <- rawToChar(cont)
df <- as.data.frame(jsonlite::fromJSON(char))

tract_mappings <- df %>%
  select(data.results.zip,
         data.results.geoid,
         data.results.res_ratio) %>%
  rename(ZIP = data.results.zip,
         TRACT = data.results.geoid,
         WEIGHT = data.results.res_ratio) %>%
  arrange(ZIP)

# add zip code to SVI using zip to tract mapping (not 1-to-1 but has weights)
svi <- svi %>%
  left_join(tract_mappings, by = "TRACT") %>% 
  distinct() %>%
  arrange(ZIP) %>%
  relocate(ZIP, .after=TRACT) %>%
  relocate(WEIGHT, .after=ZIP) %>%
  
  # remove svi data with no zip code
  filter(!is.na(ZIP))

svi_cols <- colnames(svi)[-(1:5)]
svi_weighted <- svi
svi_weighted[svi_cols] <- svi_weighted[svi_cols] * svi_weighted$WEIGHT / 100 # note: divide by 100 to convert to proportions

# Calculate sumproduct for SVI factors by zip (factor * weight)
svi_numerator <- as.data.frame(data.table::setDT(svi_weighted)[, lapply(.SD, sum), by = ZIP, .SDcols = svi_cols])

# Calculate weight totals by zip
svi_denominator <- svi %>%
  select(ZIP,
         WEIGHT) %>%
  group_by(ZIP) %>%
  summarise(Weights_Total = sum(WEIGHT))
                     
# Calculate weighted average
svi_zip <- svi_numerator
svi_zip[svi_cols] <- svi_zip[svi_cols] / svi_denominator$Weights_Total

# add SVI_ to column names
colnames(svi_zip) <- paste0("SVI_", colnames(svi_zip))
```

#####################################################################
############### Create Zip Code-Level Aggregate Table ###############
#####################################################################

```{r create zip code table}
### map ZCTA to zip code to state
# import zcta to zip crosswalk from UDS Mapper (2021 version)
zcta_to_zip_file <- readxl::read_excel("F:\\Data_Science\\SDoH\\Data\\ZIPCodetoZCTACrosswalk2021UDS.xlsx")
zcta_to_zip <- zcta_to_zip_file %>%
  mutate(ZCTA = as.numeric(ZCTA)) %>%
  select(STATE,
         ZCTA,
         ZIP_CODE)

### map state to region
state_to_region_file <- readxl::read_excel("F:\\Data_Science\\SDoH\\Data\\state to region mapping 20250121.xlsx")

table_zip <- acs %>%
  left_join(cdc, by="ZCTA") %>%
  
  # add state and zip code
  left_join(zcta_to_zip, by="ZCTA") %>%
 
  left_join(svi_zip, by=c("ZIP_CODE" = "SVI_ZIP")) %>%
  
  # add region
  left_join(state_to_region_file, by="STATE") %>%
  
  # reorder variables
  relocate(REGION,
           STATE,
           ZIP_CODE) %>%
  select(-ZCTA) %>%
  filter(!is.na(ZIP_CODE)) 

# save as RDS
# saveRDS(table_zip, file = "F:\\Data_Science\\SDoH\\Code\\SDoH\\zip_table 20240121.RDS")
```
